requirements.txt
Flask-Migrate
Flask-SQLAlchemy
python-dotenv
SQLAlchemy-Utils
WTForms-Alchemy
